Fixed two problems: auditing and identifying missing images
Issue #8 was to create a table that we can track duplicate appids from
steam.

We also added an audit table to keep track of the steam information that
we are pulling and then can add this data into our active informaiton
daily.

The check_urls.rb is a tool that tests for images and steam locations to
see if we can send the user to a steam website and load a steam image.